more git page splitting
